Override the right entry point for Active Record 8.1
Since https://github.com/rails/rails/commit/207a254cedef2c381c2898bac960b91ce14ab3a7
the implementation of what was before `_run_commit_callbacks` and
`_run_rollback_callbacks` has changed to `_run_commit_callbacks!` and
`_run_rollback_callbacks!` respectively.

When the callbacks are first called in a class, the _run_commit_callbacks
and _run_rollback_callbacks methods are defined to point to the new
implementations with the exclamation mark, and that doesn't trigger any
of the `run_#{callback}_callbacks` methods defined in the inheritance
chain. Therefore, we need to override the right methods depending on the
Active Record version.