Patch run_callbacks instead
Rather than patching the `!` version of the callbacks, we patch the `run_callbacks` method.
This way, we don't need to worry about callbacks potentially being defined _before_ we include the gem,
which can lead to Rails aliasing the non-patched version of the `_run_#{callback}_callbacks!` methods.

Plus, `run_callbacks` is public API, so it's better to patch this anyways.